requirements.txt
langchain
faiss-cpu
openai
pypdf
tiktoken
python-dotenv
sentence-transformers
transformers
torch
numpy
